requirements.txt
pandas
numpy
matplotlib
networkx
scikit-learn
sentence-transformers
pypaperbot
openai
tqdm
python-dotenv